+++ R/ggmsa.R
@@ -20,6 +20,7 @@
 ##' @param disagreement a logical value. Displays characters that disagreememt to consensus(excludes ambiguous disagreements).
 ##' @param ignore_gaps a logical value. When selected TRUE, gaps in column are treated as if that row didn't exist.
 ##' @param ref a character string. Specifying the reference sequence which should be one of input sequences when 'consensus_views' is TRUE.
+##' @param show.legend logical. Should this layer be included in the legends?
 ##' @return ggplot object
 ##' @importFrom tidyr gather
 ##' @importFrom ggplot2 ggplot
@@ -57,12 +58,12 @@
 ggmsa <- function(msa, start = NULL, end = NULL, font = "helvetical",
                   color = "Chemistry_AA", custom_color = NULL, order = NULL, char_width = 0.9, none_bg = FALSE, by_conservation = FALSE,
                   posHighligthed = NULL, seq_name = NULL, border = NULL, consensus_views = FALSE, use_dot = FALSE,
-                  disagreement = TRUE, ignore_gaps = FALSE, ref = NULL) {
+                  disagreement = TRUE, ignore_gaps = FALSE, ref = NULL, show.legend = FALSE) {
     data <- tidy_msa(msa, start = start, end = end)
 
     ggplot() + geom_msa(data, font = font, color = color, custom_color = custom_color, order = order, char_width = char_width, none_bg = none_bg,
                         by_conservation = by_conservation, posHighligthed = posHighligthed, seq_name = seq_name, border = border,
-                        consensus_views = consensus_views,use_dot = use_dot, disagreement = disagreement, ignore_gaps = ignore_gaps, ref = ref) +
+                        consensus_views = consensus_views,use_dot = use_dot, disagreement = disagreement, ignore_gaps = ignore_gaps, ref = ref, show.legend = show.legend) +
                theme_msa()
 
 }